requirements.txt
google-adk
boto3
dotenv
greenlet
aiosqlite
pydantic  